lb/requirements.txt
requests
prometheus-api-client
pandas
numpy
scikit-learn==1.3.2
joblib
python-dotenv
